Refactor pruning code to util.
