+++ packages/util/src/types.ts
@@ -2,6 +2,8 @@
 // Copyright 2021 Vulcanize, Inc.
 //
 
+import { QueryRunner } from 'typeorm';
+
 export interface BlockProgressInterface {
   id: number;
   blockHash: string;
@@ -41,7 +43,9 @@ export interface IndexerInterface {
   getBlockProgress (blockHash: string): Promise<BlockProgressInterface | undefined>
   getEvent (id: string): Promise<EventInterface | undefined>
   getSyncStatus (): Promise<SyncStatusInterface | undefined>;
+  getBlock (blockHash: string): Promise<any>
   getBlocksAtHeight (height: number, isPruned: boolean): Promise<BlockProgressInterface[]>;
+  getBlockEvents (blockHash: string): Promise<Array<EventInterface>>
   blockIsAncestor (ancestorBlockHash: string, blockHash: string, maxDepth: number): Promise<boolean>;
   updateBlockProgress (blockHash: string, lastProcessedEventIndex: number): Promise<void>
   updateSyncStatusChainHead (blockHash: string, blockNumber: number): Promise<SyncStatusInterface>
@@ -54,4 +58,14 @@ export interface EventWatcherInterface {
   getBlockProgressEventIterator (): AsyncIterator<any>
   initBlockProcessingOnCompleteHandler (): Promise<void>
   initEventProcessingOnCompleteHandler (): Promise<void>
+}
+
+export interface DatabaseInterface {
+  createTransactionRunner(): Promise<QueryRunner>;
+  getBlocksAtHeight (height: number, isPruned: boolean): Promise<BlockProgressInterface[]>;
+  getBlockProgress (blockHash: string): Promise<BlockProgressInterface | undefined>;
+  markBlockAsPruned (queryRunner: QueryRunner, block: BlockProgressInterface): Promise<BlockProgressInterface>;
+  updateSyncStatusIndexedBlock (queryRunner: QueryRunner, blockHash: string, blockNumber: number): Promise<SyncStatusInterface>;
+  updateSyncStatusChainHead (queryRunner: QueryRunner, blockHash: string, blockNumber: number): Promise<SyncStatusInterface>;
+  updateSyncStatusCanonicalBlock (queryRunner: QueryRunner, blockHash: string, blockNumber: number): Promise<SyncStatusInterface>;
 }
